Move parseCompletions to kite-data-utils, improve prefix for nested snippet display texts
