+++ lib/data-loader.js
@@ -4,7 +4,7 @@ const KiteAPI = require('kite-api');
 const KiteConnect = require('kite-connector');
 const metrics = require('./metrics');
 const {escapeHTML} = require('./highlighter');
-const {appendCompletionsFooter, symbolTypeString, completionRightLabel, kindLabel} = require('./kite-data-utils');
+const {appendCompletionsFooter, parseSnippetCompletion, symbolTypeString, kindLabel} = require('./kite-data-utils');
 
 const pendingStatusRequestByEditor = {};
 
@@ -78,50 +78,15 @@ const DataLoader = {
     .then(completions => {
       var parsedCompletions = [];
       completions.forEach(function(c) {
-        parsedCompletions.push(DataLoader.parseCompletion(editor, c, ''));
+        parsedCompletions.push(parseSnippetCompletion(editor, c, ''));
         if (c.children) {
           c.children.forEach(function(child) {
-            parsedCompletions.push(DataLoader.parseCompletion(editor, child, '...'));
+            parsedCompletions.push(parseSnippetCompletion(editor, child, String.fromCharCode(0x00A0).repeat(4)));
           });
         }
       });
       return parsedCompletions;
     });
-  },
-
-  parseCompletion(editor, c, displayPrefix) {
-    const completion = {
-      text: c.snippet.text,
-      displayText: displayPrefix + c.display,
-      className: 'kite-completion',
-    };
-    if (c.replace) {
-      completion.replacementRange = {
-        begin: editor.getBuffer().positionForCharacterIndex(c.replace.begin),
-        end: editor.getBuffer().positionForCharacterIndex(c.replace.end),
-      };
-    }
-    if (c.snippet.placeholders.length > 0) {
-      var snippetStr = c.snippet.text;
-      var offset = 0;
-      for (let i = 0; i < c.snippet.placeholders.length; i++) {
-        let placeholder = c.snippet.placeholders[i];
-        placeholder.begin += offset;
-        placeholder.end += offset;
-        snippetStr = snippetStr.slice(0, placeholder.begin)
-        + '${' + (i + 1).toString() + ':' + snippetStr.slice(placeholder.begin, placeholder.end) 
-        + '}' + snippetStr.slice(placeholder.end);
-        offset += 5;
-      }
-      completion.snippet = snippetStr;
-    }
-    completion.type = c.hint;
-    completion.rightLabelHTML = kindLabel(c.hint);
-    if (c.documentation.text) {
-      completion.descriptionHTML = appendCompletionsFooter(c, escapeHTML(c.documentation.text));
-      completion.description = c.documentation.text;
-    }
-    return completion;
   },
 
   getSignaturesAtPosition(editor, position) {
