+++ lib/kite-data-utils.js
@@ -12,7 +12,7 @@ const {
   getFunctionDetails,
   idToDottedPath,
 } = require('./utils');
-const {highlightChunk} = require('./highlighter');
+const {escapeHTML, highlightChunk} = require('./highlighter');
 
 const idIsEmpty = (id) =>
   !id || id === '';
@@ -37,7 +37,7 @@ const parameterDefault = (p) => {
     case 'python':
     case 'javascript':
       return detailNotEmpty(p, 'default_value')
-        ? `<span class="syntax--source"><span class="syntax--keyword syntax--operator syntax--assignment">=</span></span>` +
+        ? '<span class="syntax--source"><span class="syntax--keyword syntax--operator syntax--assignment">=</span></span>' +
             `<span class="parameter-default">${highlightChunk(head(detailGet(p, 'default_value')).repr)}</span>`
         : '';
     default:
@@ -286,6 +286,41 @@ const appendCompletionsFooter = (completion, descriptionHTML) => {
   </div>`;
 };
 
+const parseSnippetCompletion = (editor, c, displayPrefix) => {
+  const completion = {
+    text: c.snippet.text,
+    displayText: displayPrefix + c.display,
+    className: 'kite-completion',
+  };
+  if (c.replace) {
+    completion.replacementRange = {
+      begin: editor.getBuffer().positionForCharacterIndex(c.replace.begin),
+      end: editor.getBuffer().positionForCharacterIndex(c.replace.end),
+    };
+  }
+  if (c.snippet.placeholders.length > 0) {
+    var snippetStr = c.snippet.text;
+    var offset = 0;
+    for (let i = 0; i < c.snippet.placeholders.length; i++) {
+      let placeholder = c.snippet.placeholders[i];
+      placeholder.begin += offset;
+      placeholder.end += offset;
+      snippetStr = snippetStr.slice(0, placeholder.begin)
+      + '${' + (i + 1).toString() + ':' + snippetStr.slice(placeholder.begin, placeholder.end) 
+      + '}' + snippetStr.slice(placeholder.end);
+      offset += 5;
+    }
+    completion.snippet = snippetStr;
+  }
+  completion.type = c.hint;
+  completion.rightLabelHTML = kindLabel(c.hint);
+  if (c.documentation.text) {
+    completion.descriptionHTML = appendCompletionsFooter(c, escapeHTML(c.documentation.text));
+    completion.description = c.documentation.text;
+  }
+  return completion;
+};
+
 module.exports = {
   appendCompletionsFooter,
   callSignature,
@@ -299,6 +334,7 @@ module.exports = {
   parameterName,
   parameterType,
   parameterValue,
+  parseSnippetCompletion,
   reportFromHover,
   returnType,
   returnTypes,
