+++ lib/kite.js
@@ -431,7 +431,8 @@ const Kite = module.exports = {
       const listElement = list.suggestionListElement
         ? list.suggestionListElement
         : atom.views.getView(list);
-
+      // Override Atom's snippets package to trigger completions request
+      // on tab stop, using autocomplete-plus's findSuggestions function.
       atom.packages.activatePackage('snippets').then(snippetsPkg => {
         const snippets = snippetsPkg.mainModule;
         const safeGoToNextTabStop = snippets && snippets.goToNextTabStop;
@@ -452,6 +453,7 @@ const Kite = module.exports = {
       const safeDisplay = manager && manager.displaySuggestions;
       const safeShowOrHide = manager && manager.showOrHideSuggestionListForBufferChanges;
       const safeConfirm = manager && manager.confirm;
+      const safeReplaceTextWithMatch = manager && manager.replaceTextWithMatch;
       const safeCursorMoved = manager && manager.cursorMoved;
       const safeHide = manager && manager.hideSuggestionList;
 
@@ -467,6 +469,7 @@ const Kite = module.exports = {
           manager.displaySuggestions = safeDisplay;
           manager.showOrHideSuggestionListForBufferChanges = safeShowOrHide;
           manager.confirm = safeConfirm;
+          manager.replaceTextWithMatch = safeReplaceTextWithMatch;
           manager.cursorMoved = safeCursorMoved;
           manager.hideSuggestionList = safeHide;
         }));
@@ -570,7 +573,6 @@ const Kite = module.exports = {
             if (desc) {
               if (suggestions.length === 0) {
                 desc.style.display = 'none';
-                // TODO: Test when suggestions > 0
                 return;
               } else {
                 desc.style.display = '';
